+++ x/stakeibc/keeper/redemption_sweep.go
@@ -1,0 +1,117 @@
+package keeper
+
+import (
+	"fmt"
+
+	sdkmath "cosmossdk.io/math"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/cosmos/gogoproto/proto"
+
+	"github.com/Stride-Labs/stride/v18/utils"
+	recordstypes "github.com/Stride-Labs/stride/v18/x/records/types"
+	"github.com/Stride-Labs/stride/v18/x/stakeibc/types"
+)
+
+// Batch transfers any unbonded tokens from the delegation account to the redemption account
+func (k Keeper) SweepAllUnbondedTokensForHostZone(ctx sdk.Context, hostZone types.HostZone, epochUnbondingRecords []recordstypes.EpochUnbondingRecord) (success bool, sweepAmount sdkmath.Int) {
+	k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "Sweeping unbonded tokens"))
+
+	// Sum up all host zone unbonding records that have finished unbonding
+	totalAmtTransferToRedemptionAcct := sdkmath.ZeroInt()
+	epochUnbondingRecordIds := []uint64{}
+	for _, epochUnbondingRecord := range epochUnbondingRecords {
+
+		// Get all the unbondings associated with the epoch + host zone pair
+		hostZoneUnbonding, found := k.RecordsKeeper.GetHostZoneUnbondingByChainId(ctx, epochUnbondingRecord.EpochNumber, hostZone.ChainId)
+		if !found {
+			continue
+		}
+
+		// Get latest blockTime from light client
+		blockTime, err := k.GetLightClientTimeSafely(ctx, hostZone.ConnectionId)
+		if err != nil {
+			k.Logger(ctx).Error(fmt.Sprintf("\tCould not find blockTime for host zone %s", hostZone.ChainId))
+			continue
+		}
+
+		k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "Epoch %d - Status: %s, Amount: %v, Unbonding Time: %d, Block Time: %d",
+			epochUnbondingRecord.EpochNumber, hostZoneUnbonding.Status.String(), hostZoneUnbonding.NativeTokenAmount, hostZoneUnbonding.UnbondingTime, blockTime))
+
+		// If the unbonding period has elapsed, then we can send the ICA call to sweep this
+		//   hostZone's unbondings to the redemption account (in a batch).
+		// Verify:
+		//      1. the unbonding time is set (g.t. 0)
+		//      2. the unbonding time is less than the current block time
+		//      3. the host zone is in the EXIT_TRANSFER_QUEUE state, meaning it's ready to be transferred
+		inTransferQueue := hostZoneUnbonding.Status == recordstypes.HostZoneUnbonding_EXIT_TRANSFER_QUEUE
+		validUnbondingTime := hostZoneUnbonding.UnbondingTime > 0 && hostZoneUnbonding.UnbondingTime < blockTime
+		if inTransferQueue && validUnbondingTime {
+			k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "  %v%s included in sweep", hostZoneUnbonding.NativeTokenAmount, hostZoneUnbonding.Denom))
+
+			if err != nil {
+				errMsg := fmt.Sprintf("Could not convert native token amount to int64 | %s", err.Error())
+				k.Logger(ctx).Error(errMsg)
+				continue
+			}
+			totalAmtTransferToRedemptionAcct = totalAmtTransferToRedemptionAcct.Add(hostZoneUnbonding.NativeTokenAmount)
+			epochUnbondingRecordIds = append(epochUnbondingRecordIds, epochUnbondingRecord.EpochNumber)
+		}
+	}
+
+	// If we have any amount to sweep, then we can send the ICA call to sweep them
+	if totalAmtTransferToRedemptionAcct.LTE(sdkmath.ZeroInt()) {
+		k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "No tokens ready for sweep"))
+		return true, totalAmtTransferToRedemptionAcct
+	}
+	k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "Batch transferring %v to host zone", totalAmtTransferToRedemptionAcct))
+
+	// Get the delegation account and redemption account
+	if hostZone.DelegationIcaAddress == "" {
+		k.Logger(ctx).Error(fmt.Sprintf("Zone %s is missing a delegation address!", hostZone.ChainId))
+		return false, sdkmath.ZeroInt()
+	}
+	if hostZone.RedemptionIcaAddress == "" {
+		k.Logger(ctx).Error(fmt.Sprintf("Zone %s is missing a redemption address!", hostZone.ChainId))
+		return false, sdkmath.ZeroInt()
+	}
+
+	// Build transfer message to transfer from the delegation account to redemption account
+	sweepCoin := sdk.NewCoin(hostZone.HostDenom, totalAmtTransferToRedemptionAcct)
+	msgs := []proto.Message{
+		&banktypes.MsgSend{
+			FromAddress: hostZone.DelegationIcaAddress,
+			ToAddress:   hostZone.RedemptionIcaAddress,
+			Amount:      sdk.NewCoins(sweepCoin),
+		},
+	}
+	k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "Preparing MsgSend from Delegation Account to Redemption Account"))
+
+	// Store the epoch numbers in the callback to identify the epoch unbonding records
+	redemptionCallback := types.RedemptionCallback{
+		HostZoneId:              hostZone.ChainId,
+		EpochUnbondingRecordIds: epochUnbondingRecordIds,
+	}
+	marshalledCallbackArgs, err := k.MarshalRedemptionCallbackArgs(ctx, redemptionCallback)
+	if err != nil {
+		k.Logger(ctx).Error(err.Error())
+		return false, sdkmath.ZeroInt()
+	}
+
+	// Send the transfer ICA
+	_, err = k.SubmitTxsDayEpoch(ctx, hostZone.ConnectionId, msgs, types.ICAAccountType_DELEGATION, ICACallbackID_Redemption, marshalledCallbackArgs)
+	if err != nil {
+		k.Logger(ctx).Error(fmt.Sprintf("Failed to SubmitTxs, transfer to redemption account on %s", hostZone.ChainId))
+		return false, sdkmath.ZeroInt()
+	}
+	k.Logger(ctx).Info(utils.LogWithHostZone(hostZone.ChainId, "ICA MsgSend Successfully Sent"))
+
+	// Update the host zone unbonding records to status IN_PROGRESS
+	err = k.RecordsKeeper.SetHostZoneUnbondingStatus(ctx, hostZone.ChainId, epochUnbondingRecordIds, recordstypes.HostZoneUnbonding_EXIT_TRANSFER_IN_PROGRESS)
+	if err != nil {
+		k.Logger(ctx).Error(err.Error())
+		return false, sdkmath.ZeroInt()
+	}
+
+	return true, totalAmtTransferToRedemptionAcct
+}
